refactor(common): fold six tiny single-use *Utils classes into natural homes
Behavior-preserving continuation of the *Utils consolidation sweep; each
fold deletes a class with 1-3 caller edits:

- ArrayUtils (one method) -> CollectionUtils#toPrimitive
- HadoopInLineFSUtils -> callers use InLineFSUtils directly; buildInlineConf
  inlined into HadoopStorageConfiguration#getInline; the Hadoop-Path adapter
  becomes a private helper in InLineFileSystem, its only user
- HoodieColumnStatsIndexUtils (two thin delegates) -> HoodieMetadataWriteUtils
- InputSplitUtils -> writeString/readString become static interface methods
  on RealtimeSplit (its only caller); trivial boolean wrappers replaced with
  direct DataOutput/DataInput calls. Kept Java-8 compatible since
  hudi-hadoop-mr pins compiler release 8.
- HoodieIncrSourceCheckpointValUtils -> folded into package sibling
  checkpoint.CheckpointUtils; tests merged into TestCheckpointUtils
- ThreadUtils -> collectActiveThreads inlined into PulsarSource, its only
  caller

Verified no reflection-based use (repo-wide string-literal scan) and no
remaining references across all file types including Scala.
