+++ hudi-common/src/main/java/org/apache/hudi/common/util/CollectionUtils.java
@@ -116,6 +116,35 @@ public class CollectionUtils {
   }
 
   /**
+   * An empty immutable {@code long} array.
+   */
+  public static final long[] EMPTY_LONG_ARRAY = new long[0];
+
+  /**
+   * <p>Converts an array of object Longs to primitives.</p>
+   *
+   * <p>This method returns {@code null} for a {@code null} input array.</p>
+   *
+   * NOTE : Adapted from org.apache.commons.lang3.ArrayUtils
+   *
+   * @param array  a {@code Long} array, may be {@code null}
+   * @return a {@code long} array, {@code null} if null array input
+   * @throws NullPointerException if array content is {@code null}
+   */
+  public static long[] toPrimitive(Long[] array) {
+    if (array == null) {
+      return null;
+    } else if (array.length == 0) {
+      return EMPTY_LONG_ARRAY;
+    }
+    final long[] result = new long[array.length];
+    for (int i = 0; i < array.length; i++) {
+      result[i] = array[i].longValue();
+    }
+    return result;
+  }
+
+  /**
    * Collects provided {@link Iterator} to a {@link Stream}
    */
   public static <T> Stream<T> toStream(Iterator<T> iterator) {
