+++ hudi-hadoop-common/src/test/java/org/apache/hudi/common/fs/inline/TestInLineFileSystem.java
@@ -20,7 +20,7 @@ package org.apache.hudi.common.fs.inline;
 
 import org.apache.hudi.common.testutils.FileSystemTestUtils;
 import org.apache.hudi.common.util.collection.Pair;
-import org.apache.hudi.hadoop.fs.inline.HadoopInLineFSUtils;
+import org.apache.hudi.storage.inline.InLineFSUtils;
 import org.apache.hudi.hadoop.fs.inline.InLineFileSystem;
 import org.apache.hudi.storage.StoragePath;
 
@@ -360,12 +360,12 @@ public class TestInLineFileSystem {
       if (inputPath.toString().contains(":")) {
         scheme = inputPath.toString().split(":")[0];
       }
-      final StoragePath actualInLineFSPath = HadoopInLineFSUtils.getInlineFilePath(
+      final StoragePath actualInLineFSPath = InLineFSUtils.getInlineFilePath(
           new StoragePath(inputPath.toUri()), scheme, 10, 10);
       assertEquals(expectedInLineFSPath, actualInLineFSPath);
 
       final StoragePath actualOuterFilePath =
-          HadoopInLineFSUtils.getOuterFilePathFromInlinePath(actualInLineFSPath);
+          InLineFSUtils.getOuterFilePathFromInlinePath(actualInLineFSPath);
       assertEquals(expectedTransformedInputPath, actualOuterFilePath);
     }
   }
