+++ hudi-utilities/src/main/java/org/apache/hudi/utilities/sources/PulsarSource.java
@@ -57,7 +57,6 @@ import static org.apache.hudi.common.table.checkpoint.CheckpointUtils.createChec
 import static org.apache.hudi.common.util.ConfigUtils.checkRequiredConfigProperties;
 import static org.apache.hudi.common.util.ConfigUtils.getLongWithAltKeys;
 import static org.apache.hudi.common.util.ConfigUtils.getStringWithAltKeys;
-import static org.apache.hudi.common.util.ThreadUtils.collectActiveThreads;
 import static org.apache.hudi.utilities.config.PulsarSourceConfig.PULSAR_SOURCE_ADMIN_ENDPOINT_URL;
 import static org.apache.hudi.utilities.config.PulsarSourceConfig.PULSAR_SOURCE_MAX_RECORDS_PER_BATCH_THRESHOLD;
 import static org.apache.hudi.utilities.config.PulsarSourceConfig.PULSAR_SOURCE_OFFSET_AUTO_RESET_STRATEGY;
@@ -264,8 +263,22 @@ public class PulsarSource extends RowSource implements Closeable {
     //       properly (see above). To work this around we employ "nuclear" option of
     //       fetching all Pulsar client threads and interrupting them forcibly (to make them
     //       shutdown)
-    collectActiveThreads().stream().sequential()
+    Arrays.stream(collectActiveThreads())
         .filter(t -> t.getName().startsWith("pulsar-client-io"))
         .forEach(Thread::interrupt);
+  }
+
+  /**
+   * Fetches all active threads currently running in the JVM.
+   */
+  private static Thread[] collectActiveThreads() {
+    ThreadGroup threadGroup = Thread.currentThread().getThreadGroup();
+    while (threadGroup.getParent() != null) {
+      threadGroup = threadGroup.getParent();
+    }
+
+    Thread[] activeThreads = new Thread[threadGroup.activeCount()];
+    threadGroup.enumerate(activeThreads);
+    return activeThreads;
   }
 }
